Create sfl_dao
I have not tested it and it is likely full of bugs. The multi-dimensional struct arrays are a bit complicated and I know they're difficult to get right (I've only done them once before) ... but this is proof of concept that should be very useful.

The features of this contract are that a user must have at least 20 tokens to place vote or create a referendum to be voted on. It will verify that they still hold the tokens at the time of tally. If they don't, their vote won't count. This vote count can be voted on and changed in the contract. As can the token address that they must hold.

This should handle multiple referendums simultaneously and people should be able to vote on each of them. It also has a feature in creating the referendum that the voter must hold X token (IE: an NFT). If this feature is not used (null argument passed), it uses the default tokens that currently require >= 20.

Let me know how to proceed from here... probably debugging and writing a rudimentary web3 front end to communicate with it... or to stop because someone else is already working on one LOL...

One thing though, I don't have a lot of time to commit to helping out but I did say that I would help when I could so hopefully this is a good start and I don't know when I'll have more free time.